取名工具，地图工具

+++ 新版界面/Workfolw/ViewSet.cs
@@ -19,11 +19,11 @@ namespace RootNS.Workfolw
         /// </summary>
         /// <param name="thisWin"></param>
         /// <param name="uc"></param>
-        public static void ForViewPointX(Window thisWin, UIElement uc, int offset = 0)
+        public static void ForViewPointX(Window thisWin, UIElement uc, double offset = 0)
         {
             thisWin.Left = uc.TranslatePoint(new Point(), Gval.View.MainWindow).X + offset;
         }
-        public static void ForViewPointY(Window thisWin, UIElement uc, int offset = 0)
+        public static void ForViewPointY(Window thisWin, UIElement uc, double offset = 0)
         {
             thisWin.Top = uc.TranslatePoint(new Point(), Gval.View.MainWindow).Y + offset;
         }
